# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ExpenseCreate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "category": {
            "enum": [
              "TOLL",
              "PARKING",
              "INSURANCE",
              "PERMIT",
              "OTHER"
            ],
            "title": "Category",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id"
          }
        },
        "required": [
          "category",
          "amount"
        ],
        "title": "ExpenseCreate",
        "type": "object"
      },
      "ExpenseDecision": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "ExpenseDecision",
        "type": "object"
      },
      "ExpenseRead": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "expense_id": {
            "format": "uuid",
            "title": "Expense Id",
            "type": "string"
          },
          "fuel_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuel Id"
          },
          "maintenance_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "trip_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Label"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id"
          },
          "vehicle_reg_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Reg No"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "expense_id",
          "vehicle_id",
          "trip_id",
          "maintenance_id",
          "fuel_id",
          "category",
          "amount",
          "currency",
          "status",
          "description",
          "version"
        ],
        "title": "ExpenseRead",
        "type": "object"
      },
...
      "FuelCreate": {
        "properties": {
          "cost": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Cost"
          },
          "fuel_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuel Date"
          },
          "liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Liters"
          },
          "odometer": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odometer"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          }
        },
        "required": [
          "vehicle_id",
          "liters",
          "cost"
        ],
        "title": "FuelCreate",
        "type": "object"
      },
      "FuelRead": {
        "properties": {
          "cost": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost",
            "type": "string"
          },
          "fuel_date": {
            "format": "date",
            "title": "Fuel Date",
            "type": "string"
          },
          "fuel_id": {
            "format": "uuid",
            "title": "Fuel Id",
            "type": "string"
          },
          "liters": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Liters",
            "type": "string"
          },
          "odometer": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odometer"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          },
          "vehicle_reg_no": {
            "title": "Vehicle Reg No",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "fuel_id",
          "vehicle_id",
          "vehicle_reg_no",
          "trip_id",
          "fuel_date",
          "liters",
          "cost",
          "odometer",
          "version"
        ],
        "title": "FuelRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TransitOps API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/expenses": {
      "get": {
        "operationId": "list_expenses_api_v1_expenses_get",
        "parameters": [
          {
            "in": "query",
            "name": "vehicle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExpenseRead"
                  },
                  "title": "Response List Expenses Api V1 Expenses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Expenses",
        "tags": [
          "finance"
        ]
      },
      "post": {
        "operationId": "create_expense_api_v1_expenses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Expense",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/expenses/{expense_id}/approve": {
      "post": {
        "operationId": "approve_expense_api_v1_expenses__expense_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Expense",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/expenses/{expense_id}/reject": {
      "post": {
        "operationId": "reject_expense_api_v1_expenses__expense_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Expense",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/fuel": {
      "get": {
        "operationId": "list_fuel_api_v1_fuel_get",
        "parameters": [
          {
            "in": "query",
            "name": "vehicle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FuelRead"
                  },
                  "title": "Response List Fuel Api V1 Fuel Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Fuel",
        "tags": [
          "finance"
        ]
      },
      "post": {
        "operationId": "create_fuel_api_v1_fuel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FuelCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuelRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Fuel",
        "tags": [
          "finance"
        ]
      }
    },
...
  }
}